chore: refactor select colorscheme button
