+++ src/components/setting/appearances/NoWeebButton.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { StyleSheet, View } from 'react-native';
+
+import GenericDialog from '@components/dialogs/GenericDialog';
+
+import { SettingListItem, RenderSetting } from '../useRenderSetting';
