+++ src/components/setting/appearances/View.tsx
@@ -1,70 +1,21 @@
 import * as React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { useTranslation } from 'react-i18next';
-import {
-  Appearance,
-  StyleSheet,
-  View,
-  ColorSchemeName,
-  ScrollView,
-} from 'react-native';
+import { StyleSheet, View, ScrollView } from 'react-native';
 
 import SkinSettings from './SkinSettings';
 import { SettingListItem, RenderSetting } from '../useRenderSetting';
 import { useNoxSetting } from '@stores/useApp';
-import GenericSelectDialog from '@components/dialogs/GenericSelectDialog';
-import {
-  SelectSettingEntry,
-  dummySelectSettingEntry,
-} from '../SetttingEntries';
-import { saveColorScheme, getColorScheme } from '@utils/ChromeStorage';
+import SelectDarkMode from './SelectDarkModeButton';
 
 enum VIEW {
   HOME = 'AppearanceHome',
   SKIN = 'SkinSetting',
 }
-const ColorSchemei18n = (
-  scheme: ColorSchemeName,
-  t: (val: string) => string
-) => {
-  switch (scheme) {
-    case 'light':
-      return t('AppearanceSettings.ColorSchemeLight');
-    case 'dark':
-      return t('AppearanceSettings.ColorSchemeDark');
-    default:
-      return t('AppearanceSettings.ColorSchemeAdaptive');
-  }
-};
 
 const Stack = createNativeStackNavigator();
 
 const MainView = ({ navigation }: NoxComponent.NavigationProps) => {
-  const { t } = useTranslation();
   const playerStyle = useNoxSetting(state => state.playerStyle);
-  const [currentSelectOption, setCurrentSelectOption] = React.useState<
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    SelectSettingEntry<any>
-  >(dummySelectSettingEntry);
-  const [selectVisible, setSelectVisible] = React.useState(false);
-
-  const selectColorScheme = async () => {
-    setSelectVisible(true);
-    const options: ColorSchemeName[] = [null, 'light', 'dark'];
-    const defaultIndex = options.indexOf((await getColorScheme()) || null);
-    setCurrentSelectOption({
-      options,
-      renderOption: (option: ColorSchemeName) => ColorSchemei18n(option, t),
-      defaultIndex: defaultIndex > -1 ? defaultIndex : 0,
-      onClose: () => setSelectVisible(false),
-      onSubmit: (index: number) => {
-        setSelectVisible(false);
-        Appearance.setColorScheme(options[index]);
-        saveColorScheme(options[index]);
-      },
-      title: t('AppearanceSettings.ColorSchemeName'),
-    } as SelectSettingEntry<ColorSchemeName>);
-  };
 
   return (
     <View
@@ -98,26 +49,13 @@ const MainView = ({ navigation }: NoxComponent.NavigationProps) => {
             settingCategory: 'AppearanceSettings',
           }}
         />
+        <SelectDarkMode />
         <SettingListItem
-          settingName="ColorScheme"
-          onPress={selectColorScheme}
-          settingCategory="AppearanceSettings"
-          modifyDescription={() =>
-            t('AppearanceSettings.ColorSchemeDesc', {
-              scheme: ColorSchemei18n(Appearance.getColorScheme(), t),
-            })
-          }
+          settingName="noWeebSkins"
+          onPress={() => console.log('pressed')}
+          settingCategory="GeneralSettings"
         />
       </ScrollView>
-      <GenericSelectDialog
-        visible={selectVisible}
-        options={currentSelectOption.options}
-        renderOptionTitle={currentSelectOption.renderOption}
-        title={currentSelectOption.title}
-        defaultIndex={currentSelectOption.defaultIndex}
-        onClose={currentSelectOption.onClose}
-        onSubmit={currentSelectOption.onSubmit}
-      />
     </View>
   );
 };
